+++ models/RegistrationChecks.php
@@ -48,6 +48,11 @@ class RegistrationChecks extends Model
         $module = Yii::$app->getModule('legal');
         $rules = [];
 
+        /* If admin creates a new user, registration checks are not required. */
+        if (Yii::$app->user->isAdmin()) {
+            return $rules;
+        }
+
         if ($this->showAgeCheck()) {
             $rules[] = [['ageCheck'], 'required', 'requiredValue' => 1, 'message' => ''];
         }
@@ -174,17 +179,17 @@ class RegistrationChecks extends Model
             return false;
         }
 
-        if ($this->showTermsCheck()) {
+        if ($this->showTermsCheck() && $this->termsCheck) {
             $module->settings->user($this->user)->set(static::SETTING_KEY_TERMS, true);
             $module->settings->user($this->user)->set(static::SETTING_KEY_TERMS . 'Time', time());
         }
 
-        if ($this->showPrivacyCheck()) {
+        if ($this->showPrivacyCheck() && $this->dataPrivacyCheck) {
             $module->settings->user($this->user)->set(static::SETTING_KEY_PRIVACY, true);
             $module->settings->user($this->user)->set(static::SETTING_KEY_PRIVACY . 'Time', time());
         }
 
-        if ($this->showAgeCheck()) {
+        if ($this->showAgeCheck() && $this->ageCheck) {
             $module->settings->user($this->user)->set(static::SETTING_KEY_AGE, true);
             $module->settings->user($this->user)->set(static::SETTING_KEY_AGE . 'Time', time());
         }